# ARM Neon SIMD optimization for hashtable findBucket()