--- a/src/hashtable.c
+++ b/src/hashtable.c
@@ -727,14 +727,14 @@ static int findKeyInBucketSSE2(hashtable *ht, bucket *b, uint8_t h2, const void 
  * consider that even if they had Neon SIMD support,
  * they have 12 entries per bucket and only 32-bit scalar registers */
 
-static inline __attribute__((hot, always_inline)) int popMatchBitmask(uint64_t *mask) {
+static inline int popMatchBitmask(uint64_t *mask) {
     /* one byte (8 bits) per item - either 0x80 or 0x00 */
     int pos = __builtin_ctzll(*mask) >> 3;
     *mask &= (*mask - 1); /* clear lowest set bit */
     return pos;
 }
 
-static __attribute__((hot)) int findKeyInBucketNeon(hashtable *ht, bucket *b, uint8_t h2, const void *key, int table, int *pos_in_bucket, int *table_index) {
+static int findKeyInBucketNeon(hashtable *ht, bucket *b, uint8_t h2, const void *key, int table, int *pos_in_bucket, int *table_index) {
     const uint8x8_t hash_vector = vld1_u8(b->hashes);             /* simple load */
     const uint8x8_t h2_vector = vdup_n_u8(h2);                    /* duplicated into every byte */
     const uint8x8_t equal_mask = vceq_u8(hash_vector, h2_vector); /* compare: 8 bits per item, 0xFF or 0x00 */