--- a/src/unit/test_bitops.c
+++ b/src/unit/test_bitops.c
@@ -10,7 +10,7 @@ extern long long popcountScalar(void *s, long count);
 #if HAVE_X86_SIMD
 extern long long popcountAVX2(void *s, long count);
 #endif
-#if defined(__aarch64__)
+#if HAVE_ARM_NEON
 extern long long popcountNEON(void *s, long count);
 #endif
 
@@ -43,7 +43,7 @@ static int test_case(const char *msg, int size) {
         long long ret_avx2 = popcountAVX2(buf, size);
         TEST_ASSERT_MESSAGE(msg, expect == ret_avx2);
 #endif
-#if defined(__aarch64__)
+#if HAVE_ARM_NEON
         long long ret_neon = popcountNEON(buf, size);
         TEST_ASSERT_MESSAGE(msg, expect == ret_neon);
 #endif