@ranshid 2025-09-07
> > LGTM. I think we plan to merge it after the RC2 is cut early next week.
> 
> @madolson Oh I am sorry - I read this wrong and thought it is TODO before RC2 :(

We can probably revert this commit if we decide so